+++ IntegrationEngine.Core/Mail/MailClient.cs
@@ -1,5 +1,5 @@
-﻿using IntegrationEngine.Core.Configuration;
-using log4net;
+﻿using Common.Logging;
+using IntegrationEngine.Core.Configuration;
 using System;
 using System.Net.Mail;
 
@@ -22,11 +22,13 @@ namespace IntegrationEngine.Core.Mail
         public void Send(MailMessage mailMessage)
         {
             ConfigureSmtpClient();
-            try {
-
+            try 
+            {
                 SmtpClient.Send(mailMessage);
-            } catch (Exception exception) {
-                Log.Error("Cannot send mail message", exception);
+            } 
+            catch (Exception exception) 
+            {
+                Log.Error(exception);
             }
         }