+++ IntegrationEngine.Tests/Api/Controllers/CronTriggerControllerTest.cs
@@ -10,10 +10,10 @@ namespace IntegrationEngine.Tests.Api.Controllers
     public class CronTriggerControllerTest
     {
         [Test]
-        public void ShouldScheduleJobWhenCronTriggerIsCreated()
+        public void ShouldScheduleJobWhenCronTriggerIsCreatedWithValidCronExpression()
         {
             var subject = new CronTriggerController();
-            var cronExpression = "0 6 * * 1-5";
+            var cronExpression = "0 6 * * 1-5 ?";
             var jobType = "MyProject.MyIntegrationJob";
             var expected = new CronTrigger() {
                 JobType = jobType,
@@ -21,6 +21,7 @@ namespace IntegrationEngine.Tests.Api.Controllers
             };
             var engineScheduler = new Mock<IEngineScheduler>();
             engineScheduler.Setup(x => x.ScheduleJobWithCronTrigger(expected));
+            engineScheduler.Setup(x => x.IsJobTypeRegistered(expected.JobType)).Returns(true);
             subject.EngineScheduler = engineScheduler.Object;
             var esRepository = new Mock<ESRepository<CronTrigger>>();
             esRepository.Setup(x => x.Insert(expected)).Returns(expected);