+++ IntegrationEngine/Api/Controllers/CronTriggerController.cs
@@ -5,6 +5,7 @@ using System.Web.Http.Description;
 using IntegrationEngine.Core.Storage;
 using IntegrationEngine.Model;
 using IntegrationEngine.Scheduler;
+using System.Web.Http.Cors;
 
 namespace IntegrationEngine.Api.Controllers
 {
@@ -44,6 +45,8 @@ namespace IntegrationEngine.Api.Controllers
         {
             if (id != trigger.Id)
                 return BadRequest();
+            if (!EngineScheduler.IsJobTypeRegistered(trigger.JobType))
+                return BadRequest("Job type is invalid: " + trigger.JobType);
             if (!trigger.CronExpressionString.IsValidCronExpression())
                 return BadRequest("Cron expression is not valid: " + trigger.CronExpressionString);
             Repository.Update(trigger);
@@ -55,6 +58,8 @@ namespace IntegrationEngine.Api.Controllers
         [ResponseType(typeof(CronTrigger))]
         public IHttpActionResult PostCronTrigger(CronTrigger trigger)
         {
+            if (!EngineScheduler.IsJobTypeRegistered(trigger.JobType))
+                return BadRequest("Job type is invalid: " + trigger.JobType);
             if (!trigger.CronExpressionString.IsValidCronExpression())
                 return BadRequest("Cron expression is not valid: " + trigger.CronExpressionString);
             var triggerWithId = Repository.Insert(trigger);