+++ IntegrationEngine/Api/Startup.cs
@@ -2,7 +2,9 @@
 using Owin;
 using System.Linq;
 using System.Web.Http;
+using System.Web.Http.Cors;
 using Microsoft.Practices.Unity;
+using IntegrationEngine.Configuration;
 
 [assembly: OwinStartup(typeof(IntegrationEngine.Api.Startup))]
 
@@ -13,8 +15,11 @@ namespace IntegrationEngine.Api
         public void Configuration(IAppBuilder appBuilder)
         {
             var  config = new HttpConfiguration();
-            config.DependencyResolver = new ContainerResolver(ContainerSingleton.GetContainer());
-            config.EnableCors();
+            var container = ContainerSingleton.GetContainer();
+            config.DependencyResolver = new ContainerResolver(container);
+            var webApiConfig = container.Resolve<EngineConfiguration>().WebApi;
+            if (webApiConfig.Origins.Any())
+                config.EnableCors(new EnableCorsAttribute(string.Join(",", webApiConfig.Origins), "*", "*"));
             config.Routes.MapHttpRoute(
                 name: "DefaultApi",
                 routeTemplate: "api/{controller}/{id}",