+++ IntegrationEngine/Properties/AssemblyInfo.cs
@@ -4,4 +4,4 @@ using System.Runtime.CompilerServices;
 // Information about this assembly is defined by the following attributes.
 // Change them to the values specific to your project.
 [assembly: AssemblyTitle("IntegrationEngine")]
-[assembly: log4net.Config.XmlConfigurator(Watch = true)]
+