+++ doc/change-log.md
@@ -5,6 +5,36 @@ currentMenu: change-log
 
 # Change Log
 
+## 2.0.0-alpha5
+
+* Features
+    * Create Read-Only web app client called [InEngine.NET Dashboard](https://github.com/ethanhann/InEngine.NET-Dashboard).
+    * Enable Cors, add "Origins" Option to Web API configuration.
+    * Use Common.Logging with NLog instead of log4net.
+    * Make Common.Logging NLog adapter configurable via IntegrationEngine.json.
+    * Validate JobType classes as integration jobs before scheduling them.
+
+## 2.0.0-alpha4
+
+* Bug
+    * Use correct MySql.Data.Entity dependency in Core package.
+
+## 2.0.0-alpha3
+
+* Chores
+    * Use shared assembly to streamline package creation
+    * Remove stale NuGet dependencies
+
+## 2.0.0-alpha2
+
+* Features
+    * Return list of job keys at /api/Job.
+    * Schedule a Job with a Trigger, when the Trigger is created via the Web API.
+    * Reschedule a Job with a Trigger, when the Trigger is updated via the Web API.
+    * Validate cron expression when creating/updating a CronTrigger.
+* Chores
+    * Use Unity IoC instead of Funq.
+
 ## 2.0.0-alpha1
 
 * Features